requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
psycopg2-binary==2.9.10
pydantic==2.9.2
bcrypt==3.2.0
Faker==28.4.1
